requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
pydantic==1.9.0
pandas
scikit-learn
redis==5.0.3
apache-airflow==2.3.0  
psycopg2-binary
apache-airflow-providers-amazon
joblib